assessment_module/requirements.txt
PyYAML
pwntools